chore: Add BlockNumber index
