+++ src/chain/redb/indexes/mod.rs
@@ -1,0 +1,13 @@
+pub mod address;
+pub mod address_payment_part;
+pub mod address_stake_part;
+pub mod block_hash;
+pub mod block_number;
+pub mod tx_hash;
+
+pub use address::AddressApproxIndexTable;
+pub use address_payment_part::AddressPaymentPartApproxIndexTable;
+pub use address_stake_part::AddressStakePartApproxIndexTable;
+pub use block_hash::BlockHashApproxIndexTable;
+pub use block_number::BlockNumberApproxIndexTable;
+pub use tx_hash::TxHashApproxIndexTable;
